+++ src/mongo/db/matcher/doc_validation_error.cpp
@@ -465,7 +465,8 @@ void finishLogicalOperatorChildError(const ListOfMatchExpression* expr,
         if (ctx->haveLatestCompleteError()) {
             if (operatorsWithOrderedClauses.find(tag) != operatorsWithOrderedClauses.end()) {
                 BSONObjBuilder subBuilder = ctx->getCurrentArrayBuilder().subobjStart();
-                subBuilder.appendNumber("index", ctx->getCurrentChildIndex());
+                subBuilder.appendNumber("index",
+                                        static_cast<long long>(ctx->getCurrentChildIndex()));
                 ctx->appendLatestCompleteError(&subBuilder);
                 subBuilder.done();
             } else {
